feat: Add on conflict play config and some chores
