+++ src/core/components/animations.rs
@@ -6,7 +6,8 @@ use std::{
     time::Duration,
 };
 
-use crate::core::components::{color::Color, maths::vector::Vector};
+use crate::core::components::{color::Color};
+use crate::utils::maths::Vector;
 
 pub struct Animations {
     animations: HashMap<String, Animation>,
