+++ src/core/components/maths/mod.rs
@@ -3,7 +3,6 @@ pub mod collider;
 pub mod coordinates;
 pub mod hierarchy;
 pub mod transform;
-pub mod vector;
 
 /// `Pivot` tells where the pivot point of a component is
 pub enum Pivot {
