+++ src/core/components/maths/transform.rs
@@ -1,4 +1,5 @@
-use crate::core::components::maths::{coordinates::Coordinates, vector::Vector};
+use crate::core::components::maths::{coordinates::Coordinates};
+use crate::utils::maths::Vector;
 
 /// represents the bounds for a Transoform with min and max values
 #[derive(Default, Debug, Copy, Clone)]
